morph_failure_analysis now computes the "failure mass" of each ngram

+++ iutools-core/src/main/java/ca/pirurvik/iutools/WordInfo.java
@@ -36,7 +36,7 @@ public class WordInfo {
 	 */
 	Integer totalDecompositions = null;
 
-	public int frequency = 0;;
+	public long frequency = 0;;
 
 	public WordInfo(Long _key) {
 		this.key = _key;
